# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CityCard API",
        "version": "1.0"
    },
    "paths": {
...
        "/admin/transports/store": {
            "post": {
                "tags": [
                    "Transports"
                ],
                "summary": "Store a new transport",
                "operationId": "storeTransport",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "cities_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the transport",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Price of transport services",
                                        "type": "integer"
                                    },
                                    "cities_id": {
                                        "description": "Id of the city associated with the transport",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transport created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transport"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/transports/update/{transport}": {
            "patch": {
                "tags": [
                    "Transports"
                ],
                "summary": "Update an existing transport",
                "operationId": "updateTransport",
                "parameters": [
                    {
                        "name": "transport",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the transport",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Price of transport services",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transport updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transport"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/transports/destroy/{transport}": {
            "delete": {
                "tags": [
                    "Transports"
                ],
                "summary": "Delete an existing transport",
                "operationId": "deleteTransport",
                "parameters": [
                    {
                        "name": "transport",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Transport delete successfully"
                    }
                }
            }
        },
        "/user/card": {
            "get": {
                "tags": [
                    "Cards"
                ],
                "summary": "Get card details",
                "operationId": "getDetailsOfCard",
                "parameters": [
                    {
                        "name": "cardId",
                        "in": "query",
                        "description": "ID of the card to retrieve details for",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful get card details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "card": {
                                            "properties": {
                                                "number": {
                                                    "type": "string"
                                                },
                                                "balance": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "tickets": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "tripHistory": {
                                                                "type": "array",
                                                                "items": {
                                                                    "$ref": "#/components/schemas/TripHistory"
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "topUpHistory": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TopUpHistory"
                                            }
                                        },
                                        "costHistory": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CostHistory"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Transport": {
                "title": "Transport",
                "required": [
                    "id",
                    "cities_id",
                    "name",
                    "price"
                ],
                "properties": {
                    "id": {
                        "description": "Ticket ID",
                        "type": "integer"
                    },
                    "cities_id": {
                        "description": "ID of the city associated with the transport",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Transport name",
                        "type": "string"
                    },
                    "price": {
                        "description": "Price of transport services",
                        "type": "number",
                        "format": "double"
                    },
                    "created_at": {
                        "description": "Timestamp when the ticket was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp when the ticket was updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "TripHistory": {
                "title": "TripHistory",
                "required": [
                    "id",
                    "ticket_id",
                    "route",
                    "date"
                ],
                "properties": {
                    "id": {
                        "description": "Ticket ID",
                        "type": "integer"
                    },
                    "ticket_id": {
                        "description": "ID of the ticket associated with the TripHistory",
                        "type": "integer"
                    },
                    "route": {
                        "description": "route",
                        "type": "string"
                    },
                    "date": {
                        "description": "Start of the route",
                        "type": "date-time"
                    },
                    "created_at": {
                        "description": "Timestamp when the ticket was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp when the ticket was updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "CostHistory": {
                "title": "CostHistory",
                "required": [
                    "id",
                    "card_id",
                    "expense",
                    "date"
                ],
                "properties": {
                    "id": {
                        "description": "Ticket ID",
                        "type": "integer"
                    },
                    "card_id": {
                        "description": "ID of the card associated with the CostHistory",
                        "type": "integer"
                    },
                    "expense": {
                        "description": "expense",
                        "type": "number",
                        "format": "double"
                    },
                    "date": {
                        "description": "expense",
                        "type": "date-time"
                    },
                    "created_at": {
                        "description": "Timestamp when the ticket was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp when the ticket was updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "TopUpHistory": {
                "title": "TopUpHistory",
                "required": [
                    "id",
                    "card_id",
                    "replenishment",
                    "date"
                ],
                "properties": {
                    "id": {
                        "description": "Ticket ID",
                        "type": "integer"
                    },
                    "card_id": {
                        "description": "ID of the card associated with the TopUpHistory",
                        "type": "integer"
                    },
                    "replenishment": {
                        "description": "replenishment",
                        "type": "number",
                        "format": "double"
                    },
                    "date": {
                        "description": "topUp",
                        "type": "date-time"
                    },
                    "created_at": {
                        "description": "Timestamp when the ticket was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp when the ticket was updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}